+++ packages/components/molecules/f-mega-modal/vue.config.js
@@ -5,6 +5,19 @@ const sassOptions = require('../../../../config/sassOptions')(rootDir);
 
 // vue.config.js
 module.exports = {
+    configureWebpack: {
+        optimization: {
+            splitChunks: {
+                cacheGroups: {
+                    node_vendors: {
+                        test: /[\\/]node_modules[\\/]/,
+                        chunks: 'all',
+                        priority: 1
+                    }
+                }
+            }
+        }
+    },
     chainWebpack: config => {
         config.module
             .rule('scss-importer')